# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/media": {
      "get": {
        "operationId": "MediaController_list",
        "parameters": [
          {
            "name": "mimePrefix",
            "required": false,
            "in": "query",
            "schema": {
              "example": "image/"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Media"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem Details error response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "admin/media"
        ]
      }
    },
    "/api/v1/admin/media/upload-session": {
      "post": {
        "operationId": "MediaController_session",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UploadSession"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem Details error response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "admin/media"
        ]
      }
    },
    "/api/v1/admin/media/{id}/replace-session": {
      "post": {
        "operationId": "MediaController_replaceSession",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UploadSession"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem Details error response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "admin/media"
        ]
      }
    },
    "/api/v1/admin/media/finalize": {
      "post": {
        "operationId": "MediaController_finalize",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Media"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem Details error response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "admin/media"
        ]
      }
    },
    "/api/v1/admin/media/{id}": {
      "patch": {
        "operationId": "MediaController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Media"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem Details error response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "admin/media"
        ]
      },
      "get": {
        "operationId": "MediaController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Media"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem Details error response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "admin/media"
        ]
      }
    },
    "/api/v1/admin/media/{id}/usages": {
      "get": {
        "operationId": "MediaController_usages",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MediaUsage"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem Details error response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "admin/media"
        ]
      }
    },
    "/api/v1/admin/media/{id}/archive": {
      "post": {
        "operationId": "MediaController_archive",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Media"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem Details error response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "admin/media"
        ]
      }
    },
    "/api/v1/admin/media/{id}/retry": {
      "post": {
        "operationId": "MediaController_retry",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Media"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem Details error response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "admin/media"
        ]
      }
    },
    "/api/v1/admin/media-folders": {
      "get": {
        "operationId": "MediaController_folders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MediaFolder"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem Details error response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "admin/media"
        ]
      },
      "post": {
        "operationId": "MediaController_createFolder",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/MediaFolder"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem Details error response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "admin/media"
        ]
      }
    },
    "/api/v1/admin/media-folders/{id}": {
      "patch": {
        "operationId": "MediaController_updateFolder",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/MediaFolder"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem Details error response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "admin/media"
        ]
      }
    },
...
  },
  "info": {
    "title": "GATEVIA API",
    "description": "GATEVIA V1 public and administration REST API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gatevia_session"
      }
    },
    "schemas": {
...
      "PaginationMeta": {
        "type": "object",
        "required": [
          "page",
          "pageSize",
          "total",
          "pageCount"
        ],
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "pageCount": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "MediaTranslation": {
        "type": "object",
        "required": [
          "locale"
        ],
        "properties": {
          "locale": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "altText": {
            "type": "string"
          },
          "caption": {
            "type": "string"
          },
          "decorative": {
            "type": "boolean"
          }
        }
      },
      "MediaVariant": {
        "type": "object",
        "required": [
          "variantKey"
        ],
        "properties": {
          "variantKey": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "width": {
            "type": "integer"
          },
          "height": {
            "type": "integer"
          },
          "mimeType": {
            "type": "string"
          }
        }
      },
      "MediaFolder": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "parentId": {
            "oneOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "Media": {
        "type": "object",
        "required": [
          "id",
          "originalFilename",
          "mimeType",
          "status",
          "sizeBytes"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "originalFilename": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending_upload",
              "processing",
              "ready",
              "failed",
              "archived"
            ]
          },
          "sizeBytes": {
            "oneOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ]
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "width": {
            "type": "integer"
          },
          "height": {
            "type": "integer"
          },
          "folderId": {
            "oneOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "folder": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/MediaFolder"
              },
              {
                "type": "null"
              }
            ]
          },
          "translations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaTranslation"
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaVariant"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "uploadedById": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "MediaUsage": {
        "type": "object",
        "required": [
          "type",
          "id"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        }
      },
      "UploadSession": {
        "type": "object",
        "required": [
          "url",
          "headers",
          "uploadToken"
        ],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "uploadToken": {
            "type": "string"
          }
        }
      },
...
    }
  }
}